+++ packages/angular/src/index.ts
@@ -1,2 +1,2 @@
+export * from './data';
 export * from './angular-providers.builder';
-export * from './http-request.builder';