+++ packages/core/src/repository/data-source/in-memory.data-source.ts
@@ -14,7 +14,7 @@ export class InMemoryDataSource<T> implements DataSource<T> {
     private objects: Map<string, T> = new Map();
     private arrays: Map<string, T[]> = new Map();
 
-    constructor(private readonly logger: Logger = new DeviceConsoleLogger()) {}
+    constructor(private readonly logger: Logger = new DeviceConsoleLogger(undefined, 'InMemoryDataSource')) {}
 
     public async get(query: Query): Promise<T> {
         if (query instanceof KeyQuery) {
@@ -67,6 +67,9 @@ export class InMemoryDataSource<T> implements DataSource<T> {
         }
 
         if (query instanceof KeyQuery) {
+            if (!query.key) {
+                this.logger.warn('key is empty');
+            }
             this.objects.set(query.key, value);
             return value;
         } else {